+++ src/main/java/com/atlauncher/gui/card/InstanceCard.java
@@ -19,7 +19,9 @@
 package com.atlauncher.gui.card;
 
 import com.atlauncher.App;
+import com.atlauncher.Gsons;
 import com.atlauncher.LogManager;
+import com.atlauncher.data.DisableableMod;
 import com.atlauncher.data.Instance;
 import com.atlauncher.data.Language;
 import com.atlauncher.evnt.listener.RelocalizationListener;
@@ -31,6 +33,7 @@ import com.atlauncher.gui.dialogs.EditModsDialog;
 import com.atlauncher.gui.dialogs.InstanceInstallerDialog;
 import com.atlauncher.gui.dialogs.ProgressDialog;
 import com.atlauncher.gui.dialogs.RenameInstanceDialog;
+import com.atlauncher.utils.Base64;
 import com.atlauncher.utils.Utils;
 
 import javax.swing.BorderFactory;
@@ -52,6 +55,9 @@ import java.awt.Cursor;
 import java.awt.Dialog.ModalityType;
 import java.awt.Dimension;
 import java.awt.FlowLayout;
+import java.awt.Toolkit;
+import java.awt.datatransfer.Clipboard;
+import java.awt.datatransfer.StringSelection;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.MouseAdapter;
@@ -61,6 +67,7 @@ import java.awt.event.WindowEvent;
 import java.io.File;
 import java.io.IOException;
 import java.sql.Timestamp;
+import java.util.ArrayList;
 import java.util.Date;
 
 /**
@@ -417,8 +424,8 @@ public class InstanceCard extends CollapsiblePanel implements RelocalizationList
                                 String[] optionss = {Language.INSTANCE.localize("common.ok")};
                                 JOptionPane.showOptionDialog(App.settings.getParent(), Language.INSTANCE.localize
                                         ("instance.cantupdate"), Language.INSTANCE.localize("instance" + "" +
-                                                ".noaccountselected"), JOptionPane.DEFAULT_OPTION, JOptionPane
-                                        .ERROR_MESSAGE, null, optionss, optionss[0]);
+                                        ".noaccountselected"), JOptionPane.DEFAULT_OPTION, JOptionPane.ERROR_MESSAGE,
+                                        null, optionss, optionss[0]);
                             } else {
                                 new InstanceInstallerDialog(instance, true, false);
                             }
@@ -445,11 +452,18 @@ public class InstanceCard extends CollapsiblePanel implements RelocalizationList
                     }
                 } else if (e.getButton() == MouseEvent.BUTTON3) {
                     JPopupMenu rightClickMenu = new JPopupMenu();
+
                     JMenuItem changeImageItem = new JMenuItem(Language.INSTANCE.localize("instance.changeimage"));
-                    JMenuItem updateItem = new JMenuItem("Update Instance");
                     rightClickMenu.add(changeImageItem);
+
+                    JMenuItem shareCodeItem = new JMenuItem(Language.INSTANCE.localize("instance.sharecode"));
+                    rightClickMenu.add(shareCodeItem);
+
+                    JMenuItem updateItem = new JMenuItem(Language.INSTANCE.localize("common.update"));
                     rightClickMenu.add(updateItem);
+
                     rightClickMenu.show(image, e.getX(), e.getY());
+
                     changeImageItem.addActionListener(new ActionListener() {
                         @Override
                         public void actionPerformed(ActionEvent e) {
@@ -472,6 +486,7 @@ public class InstanceCard extends CollapsiblePanel implements RelocalizationList
                             }
                         }
                     });
+
                     updateItem.addActionListener(new ActionListener() {
                         @Override
                         public void actionPerformed(ActionEvent e) {
@@ -481,11 +496,11 @@ public class InstanceCard extends CollapsiblePanel implements RelocalizationList
                                         .localize("common.no"), Language.INSTANCE.localize("instance" + "" +
                                         ".dontremindmeagain")};
                                 int ret = JOptionPane.showOptionDialog(App.settings.getParent(), "<html><p " +
-                                                "align=\"center\">" + Language.INSTANCE.localize("instance" + "" +
-                                                ".updatenow", "<br/><br/>") + "</p></html>", Language.INSTANCE
-                                        .localize("instance" + "" +
-                                                ".updateavailable"), JOptionPane.DEFAULT_OPTION, JOptionPane
-                                        .ERROR_MESSAGE, null, options, options[0]);
+                                        "align=\"center\">" + Language.INSTANCE.localize("instance" + "" +
+                                        ".updatenow", "<br/><br/>") + "</p></html>", Language.INSTANCE.localize
+                                        ("instance" + "" +
+                                        ".updateavailable"), JOptionPane.DEFAULT_OPTION, JOptionPane.ERROR_MESSAGE,
+                                        null, options, options[0]);
                                 if (ret == 0) {
                                     if (App.settings.getAccount() == null) {
                                         String[] optionss = {Language.INSTANCE.localize("common.ok")};
@@ -516,6 +531,35 @@ public class InstanceCard extends CollapsiblePanel implements RelocalizationList
                                         App.settings.setMinecraftLaunched(true);
                                     }
                                 }
+                            }
+                        }
+                    });
+
+                    shareCodeItem.addActionListener(new ActionListener() {
+                        @Override
+                        public void actionPerformed(ActionEvent e) {
+                            ArrayList<String> installedOptionalMods = new ArrayList<String>();
+
+                            for (DisableableMod mod : instance.getInstalledMods()) {
+                                if (mod.isOptional() && !mod.isUserAdded()) {
+                                    installedOptionalMods.add(mod.getName());
+                                }
+                            }
+
+                            if (installedOptionalMods.size() > 0) {
+                                App.TOASTER.pop(Language.INSTANCE.localize("instance.sharecodecopied"));
+                                LogManager.info("Share code copied to clipboard");
+
+                                String code = Gsons.DEFAULT.toJson(installedOptionalMods);
+
+                                code = Base64.encodeBytes(code.getBytes());
+
+                                StringSelection text = new StringSelection(code);
+                                Clipboard clipboard = Toolkit.getDefaultToolkit().getSystemClipboard();
+                                clipboard.setContents(text, null);
+                            } else {
+                                App.TOASTER.pop(Language.INSTANCE.localize("instance.nooptionalmods"));
+                                LogManager.info("Share code copied to clipboard");
                             }
                         }
                     });